+++ app/components/candidate_interface/monthly_statistics_table_component.html.erb
@@ -21,7 +21,6 @@
         <% end %>
       </tbody>
 
-      <!--  Still need to work out these totals-->
       <tfoot>
       <tr class="govuk-table__row">
         <th scope="row" class="govuk-table__header">Total</th>
